third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/semantics/interactive-elements/the-dialog-element/modal-dialog-scroll-height-expected.txt

This is a testharness.js-based test.
[FAIL] dialogs should be centered before computing overflow.
  assert_equals: expected 493 but got 500
Harness: the test ran to completion.
